split examples spec into two
